[fixtures] refactoring: make fixtures loader an interface
We need to add support of other databases (mysql, mongodb, etc)

Changes:
 - postgres loader extracted to own package
 - added common interface instead of specific loader
 - added param db-type to specify the database type
    (only postgres is supported now)
 - test-cases definition extracted to separate files from code
    (for more clarity and for shared access between tests for other DBs)
 - added param DbType to RunWithTestingParams
